webkit/data/layout_tests/pending/fast/events/special-key-events-in-input-text.html

<p>To test manually, press keys and compare results to other browsers.</p>
<input type="text"></input>
<div id="log"></div>

<script>
function log(msg) {
   document.getElementById("log").innerHTML+= msg + "<br />";
}

function logEvent(e) {
   var target = e.target;
   var type = e.type;

   var info = target.tagName + " - " + type + ' - ';
   if (type == "textInput") {
     info += e.data;
   } else if (type == "keydown" || type == "keypress" || type == "keyup") {
     info += [e.ctrlKey, e.altKey, e.shiftKey, e.metaKey] +
         ' - ' + e.keyIdentifier +
         ' - ' + e.keyCode +
         ' - ' + e.charCode;
   }
   info += '. Value: "' + target.value + '".';
   log(info);
}

log("target - type - " + ["ctrlKey", "altKey", "shiftKey", "metaKey"]
       + ' - ' + "keyIdentifier"
       + ' - ' + "keyCode"
       + ' - ' + "charCode");

var input = document.getElementsByTagName("input")[0];
input.addEventListener('textInput', logEvent, false);
input.addEventListener('keypress', logEvent, false);
input.addEventListener('keydown', logEvent, false);
input.addEventListener('keyup', logEvent, false);
input.focus();

if (window.layoutTestController) {
   layoutTestController.dumpAsText();
   log("Ampersand:");
   eventSender.keyDown("7", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
   log("Percent:");
   eventSender.keyDown("5", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
   log("Left parentheses:");
   eventSender.keyDown("9", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
   log("Right parentheses:");
   eventSender.keyDown("0", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
   log("Left Curly Brace:");
   eventSender.keyDown("\xDB", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
   log("Right Curly Brace:");
   eventSender.keyDown("\xDD", ["shiftKey"]);
   log("Backspace:");
   eventSender.keyDown("\x08", []);
}
</script>
