webkit/data/layout_tests/pending/fast/events/special-key-events-in-input-text-expected.txt

To test manually, press keys and compare results to other browsers.


target - type - ctrlKey,altKey,shiftKey,metaKey - keyIdentifier - keyCode - charCode
Ampersand:
INPUT - keydown - false,false,true,false - U+0037 - 55 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 38 - 38. Value: "".
INPUT - textInput - &. Value: "".
INPUT - keyup - false,false,true,false - U+0037 - 55 - 0. Value: "&".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: "&".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
Percent:
INPUT - keydown - false,false,true,false - U+0035 - 53 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 37 - 37. Value: "".
INPUT - textInput - %. Value: "".
INPUT - keyup - false,false,true,false - U+0035 - 53 - 0. Value: "%".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: "%".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
Left parentheses:
INPUT - keydown - false,false,true,false - U+0039 - 57 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 40 - 40. Value: "".
INPUT - textInput - (. Value: "".
INPUT - keyup - false,false,true,false - U+0039 - 57 - 0. Value: "(".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: "(".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
Right parentheses:
INPUT - keydown - false,false,true,false - U+0030 - 48 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 41 - 41. Value: "".
INPUT - textInput - ). Value: "".
INPUT - keyup - false,false,true,false - U+0030 - 48 - 0. Value: ")".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: ")".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
Left Curly Brace:
INPUT - keydown - false,false,true,false - U+00DB - 219 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 123 - 123. Value: "".
INPUT - textInput - {. Value: "".
INPUT - keyup - false,false,true,false - U+00DB - 219 - 0. Value: "{".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: "{".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
Right Curly Brace:
INPUT - keydown - false,false,true,false - U+00DD - 221 - 0. Value: "".
INPUT - keypress - false,false,true,false - - 125 - 125. Value: "".
INPUT - textInput - }. Value: "".
INPUT - keyup - false,false,true,false - U+00DD - 221 - 0. Value: "}".
Backspace:
INPUT - keydown - false,false,false,false - U+0008 - 8 - 0. Value: "}".
INPUT - keyup - false,false,false,false - U+0008 - 8 - 0. Value: "".
